+++ presentation/src/main/java/com/threegap/bitnagil/presentation/reporthistory/contract/ReportHistorySideEffect.kt
@@ -1,0 +1,3 @@
+package com.threegap.bitnagil.presentation.reporthistory.contract
+
+sealed interface ReportHistorySideEffect
