--- presentation/src/main/java/com/threegap/bitnagil/presentation/reporthistory/model/ReportHistorySideEffect.kt
@@ -1,3 +1,0 @@
-package com.threegap.bitnagil.presentation.reporthistory.model
-
-sealed interface ReportHistorySideEffect
